feat: 모델 클래스 작성 및 오류 수정
