+++ lib/data/data_source/recipe_data_source_impl.dart
@@ -1,10 +1,11 @@
 import 'dart:convert';
 
 import 'package:flutter_recipe_app/data/data_source/recipe_data_source.dart';
-import 'package:flutter_recipe_app/data/dto/recipe_dto.dart';
 
 import '../../core/routing/response.dart';
 import 'package:http/http.dart' as http;
+
+import '../dto/recipes_dto.dart';
 
 class RecipeDataSourceImpl implements RecipeDataSource {
   final _baseUrl =
@@ -16,7 +17,7 @@ class RecipeDataSourceImpl implements RecipeDataSource {
     : _client = client ?? http.Client();
 
   @override
-  Future<Response<RecipeDto>> getRecipes() async {
+  Future<Response<RecipesDto>> getRecipes() async {
     final response = await _client.get(Uri.parse(_baseUrl));
 
     final jsonMap = jsonDecode(response.body) as Map<String, dynamic>;
@@ -24,7 +25,7 @@ class RecipeDataSourceImpl implements RecipeDataSource {
     return Response(
       statusCode: response.statusCode,
       header: response.headers,
-      body: RecipeDto.fromJson(jsonMap), // RecipesResultDto
+      body: RecipesDto.fromJson(jsonMap), // RecipesResultDto
     );
   }
 }
