--- lib/data/model/ingredient.dart
@@ -1,12 +1,0 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'ingredient.freezed.dart';
-
-@freezed
-abstract class Ingredient with _$Ingredient {
-  const factory Ingredient({
-    required int id,
-    required String name,
-    required String imageUrl,
-  }) = _Ingredient;
-}
