+++ lib/data/model/ingredients.dart
@@ -1,13 +1,14 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'ingredient.dart';
+import 'recipes_ingredients.dart';
 
 part 'ingredients.freezed.dart';
 
 @freezed
 abstract class Ingredients with _$Ingredients {
   const factory Ingredients({
-    required Ingredient ingredient,
-    required int amount,
+    required int id,
+    required String name,
+    required String image,
   }) = _Ingredients;
 }
