+++ lib/presentation/component/card/recipe_card.dart
@@ -144,7 +144,7 @@ class RecipeCard extends StatelessWidget {
                                       shape: BoxShape.circle,
                                     ),
                                     child: Icon(
-                                      recipe.isBookmarked ? Icons.bookmark : Icons.bookmark_border, // 修正
+                                      Icons.bookmark, // 修正
                                       color: Colors.white,
                                       size: 15,
                                     ),
